Create an EditTask Get and Post
Basically the same as an Edit action, but the when coming back, we come
back to the parent task and not the list.  Note that this whole idea may
need to be expanded to be the general Edit paradigm for work items,
which means that the controller will no longer be able to inherit from
ReadWriteController (which itself should probably be renamed to
SimpleReadWriteController).