Restore bi-directional mapping for work items and parent work items.
This was previously removed because I didn't really need it and it was causing problems.
However, I had a situation where I required the bi-directional mapping between parent and child,
so I researched how to get around the issue.  Now thtat the JSON converter is globally configured
to get around the problem, I can restore the more proper mapping.