Rework the editor logic.
This involved stripping out much of the Editor JavaScript logic, to the
point that there is no longer an EditorBase anymore.  The controller
logic was alo changed to return to the caller instead of the editor
upon a successful save.

Basically, this is a complete revert of the functionallity added a couple
of versions ago, back to the original behavior, which I like better.