Add a details view between the index and editor.
This makes the User's controller consistent with the admin controllers.

I like the styling of this one and I think I will copy it for the other
details views.

NOTE: as part of this, I also created a base class for the controllers
that will allow the navigation stack to be handled in a consistent
manner.