Create a details view for the work items.
This works the same as the admin ones created before with the exception
that work items can have a task list depending on their type.

The task list is displayed for those types and hidden for the others.
The user is able to manipulate the task list from the details view.