Create fallback editors.
These editors will be used when client side validation was not able to be
done and the validation failed on the server side.  When we come back from
the POST, we need something simple to load, and this is the simplest
solution.

The most common errors are ones that can be enforced on the client first,
so these editors should rarely be used.