Remove error messages from data service.
The error messages are a UI concern and thus have been moved to the web project.
The only exception is he error message for duplicate names.  That condition can only be
effectively checked in the data service, and thus remains there.

Many of the other errors can be enforced on the client, but only if the error actually is
expressed on the view model, which is another reason for moving these messages and
associated attributes.